+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Support Marketing API version 1.3.0.
 * Support Marketing API version 1.2.0.
 * Support Order API version v1_beta.6.0.
+* Support Trading API version 1019.
 * Support Trading API version 1007.
 * Support Shopping API version 1007.
 * Support Browse API version v1_beta.7.0.